Generalise list matchers to accept any Foldable instance.
This relies on the updated definitions of various functions in the base-4.8.0.0
Prelude, so have updated dependencies and bumped version number.  It may be
possible to revert to base-4.5.0.0 if desired by explicitly using Foldable
functions rather than defaulting to list functions, but as I don't have an
appropriate installation to test it on I've not done so.